pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "guardrails-service"
version = "0.1.0"
description = "Extendable guardrails service for masking sensitive data in LLM traffic"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Antimatter Guardrails" }]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "PyYAML>=6.0.1",
  "redis>=5.0.7",
  "orjson>=3.10.6",
  "regex>=2024.5.15",
  "transformers>=4.55.0",
  "torch>=2.3.0",
  "huggingface_hub>=0.25.0",
  "phonenumbers>=8.13.0",
  "natasha>=1.6.0",
  "presidio-analyzer>=2.2.357",
]

[project.optional-dependencies]
cuda = [
  "nvidia-pytriton>=0.7.0",
]
eval = [
  "datasets>=3.0.0",
]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "respx>=0.21.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
